ДЗ№6. ATM
