+++ hw06-ATM/src/main/java/ru/otus/hw06/support/BanknoteCell.java
@@ -1,59 +1,45 @@
 package ru.otus.hw06.support;
 
-import ru.otus.hw06.exceptions.IllegalOperationException;
-import ru.otus.hw06.exceptions.NotEnoughBanknotesException;
+/*
+Ячейка с банкнотами
+ */
 
-public class BanknoteCell implements BanknoteCellInterface {
+public interface BanknoteCell {
+    /*
+    * Получение денег (банкнот) в определенном количестве
+    *
+    * @param value - количество банкнот, которые необходимо выдать
+    * @return количество банконт, которые выдаются из ячейки
+    */
+    Integer getBanknotes(int value);
 
-    private final Banknote banknoteType;
-    private Integer countBanknoteInCell;
+    /*
+    * Добавление денег в хранилище в определенном размере
+    *
+    * @param value - количество банкнот, которые необходимо положить в ячейку
+    * @param banknote номинал банкноты {@link Banknote}
+    */
+    void putBanknotes(int value);
 
-    //Создаем ячейку с номиналом и количеством банкнот
-    public BanknoteCell(Banknote banknote, Integer amount) {
-        this.countBanknoteInCell = amount;
-        this.banknoteType = banknote;
-    }
+    /*
+    * какие банкноты лежат в ячейке
+    *
+    * @return тип банкнот в ячейке {@link Banknote}
+    */
+    Banknote getBanknotesType();
 
-    //Получение денег (банкнот) в определенном количестве
-    /** {@inheritDoc} */
-    @Override
-    public Integer getBanknotes(int value) {
-        if (value<0){
-            throw new IllegalOperationException("Невозможно выдать отрицательную сумму");
-        }
-        if (this.countBanknoteInCell - value < 0) {
-            throw new NotEnoughBanknotesException(value, this.countBanknoteInCell);
-        }
-        this.countBanknoteInCell -= value;
-        return value;
-    }
+    /*
+    * количество банкнот в ячейке
+    *
+    * @return доступное количество банконт в ячейке
+    */
+    Integer getAvailableAmountInCell();
 
-    //Добавление денег в хранилище в определенном размере
-    /** {@inheritDoc} */
-    @Override
-    public void putBanknotes(int value) {
-        if (value<0){
-            throw new IllegalOperationException("Невозможно положить отрицательную сумму");
-        }
-        this.countBanknoteInCell += value;
-    }
+    /*
+    * колчество денег в ячейке
+    *
+    * @return доступное количество денег в ячейке (количество банкнот * номинал банкноты)
+    */
+    Integer getAvailableAmountNominalInCell();
 
-    //какие банкноты лежат в ячейке
-    /** {@inheritDoc} */
-    @Override
-    public Banknote getBanknotesType() { return this.banknoteType; }
-
-    //количество банкнот в ячейке
-    /** {@inheritDoc} */
-    @Override
-    public Integer getAvailableAmountInCell(){
-        return this.countBanknoteInCell;
-    }
-
-    //колчество денег в ячейке
-    /** {@inheritDoc} */
-    @Override
-    public Integer getAvailableAmountNominalInCell(){
-        return this.banknoteType.getNominal()*this.countBanknoteInCell;
-    }
 }
